+++ src/main/java/com/revolvingmadness/sculk/language/interpreter/VariableTable.java
@@ -5,8 +5,10 @@ import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.instances.*;
 import com.revolvingmadness.sculk.language.builtins.classes.types.*;
-import com.revolvingmadness.sculk.language.builtins.functions.io.PrintFunction;
-import com.revolvingmadness.sculk.language.builtins.functions.types.TypeFunction;
+import com.revolvingmadness.sculk.language.builtins.functions.Base64DecodeFunction;
+import com.revolvingmadness.sculk.language.builtins.functions.Base64EncodeFunction;
+import com.revolvingmadness.sculk.language.builtins.functions.PrintFunction;
+import com.revolvingmadness.sculk.language.builtins.functions.TypeFunction;
 import com.revolvingmadness.sculk.language.lexer.TokenType;
 import net.minecraft.world.World;
 
@@ -84,6 +86,8 @@ public class VariableTable {
     }
 
     private void declareFunctions() {
+        this.declare(List.of(TokenType.CONST), "base64decode", new Base64DecodeFunction());
+        this.declare(List.of(TokenType.CONST), "base64encode", new Base64EncodeFunction());
         this.declare(List.of(TokenType.CONST), "print", new PrintFunction());
         this.declare(List.of(TokenType.CONST), "type", new TypeFunction());
     }